feat: install redirects

+++ zhiva/src/cmd/install.ts
@@ -21,14 +21,63 @@ if (existsSync(name)) {
     process.chdir(name);
     await $`git pull`;
 } else {
-    await $`git clone https://github.com/${name}.git ${name}`;
+    let branch: string | undefined = undefined;
+    let cloneName = name;
+
+    try {
+        let configBranch = "HEAD";
+        let configPath = "zhiva.json";
+        const maxRedirects = 15;
+
+        for (let i = 0; i < maxRedirects; i++) {
+            if (i === maxRedirects - 1) {
+                console.error("Exceeded max redirects while resolving config.");
+                break;
+            };
+
+            const res = await fetch(`https://raw.githubusercontent.com/${cloneName}/${configBranch}/${configPath}`);
+            if (!res.ok) break;
+            const tempConfig = await res.json();
+
+            if (tempConfig.redirect_repo) {
+                console.log(`Redirecting from "${cloneName}" to "${tempConfig.redirect_repo}"...`);
+                name = tempConfig.redirect_repo;
+                cloneName = tempConfig.redirect_repo;
+                configBranch = "HEAD";
+                configPath = "zhiva.json";
+                continue;
+            }
+
+            if (tempConfig.redirect_zhiva) {
+                const [newPath, newBranch] = tempConfig.redirect_zhiva.split("#");
+                configPath = newPath.startsWith("./") ? newPath.slice(2) : newPath;
+                if (newBranch) configBranch = newBranch;
+                console.log(`Redirecting zhiva.json to "${tempConfig.redirect_zhiva}"...`);
+                continue;
+            }
+
+            branch = tempConfig.branch;
+            break;
+        }
+    } catch (e) {
+        console.error("Error resolving zhiva config:", e.message);
+    }
+
+    const cloneUrl = `https://github.com/${cloneName}.git`;
+    if (branch) {
+        await $`git clone -b ${branch} ${cloneUrl} ${name}`;
+    } else {
+        await $`git clone ${cloneUrl} ${name}`;
+    }
     process.chdir(name);
 }
 
-await $`bun install --production --force`;
-const pkg = JSON.parse(readFileSync("package.json", "utf-8"));
-if (pkg.scripts?.build) {
-    await $`bun run build`;
+if (existsSync("package.json")) {
+    await $`bun install --production --force`;
+    const pkg = JSON.parse(readFileSync("package.json", "utf-8"));
+    if (pkg.scripts?.build) {
+        await $`bun run build`;
+    }
 }
 
 const updated = await db.updateOneOrAdd("apps", { name }, { updatedAt: Date.now() });
